feat: adds `mix check` github action
`mix check` is a mix task made available by the ex_check dependency.
It has the following optional dependencies:
- credo
- dialyxir
- doctor
- ex_doc
- sobelow
- mix_audit

If the dependences are installed in the project then they will be run as part
of `mix check`.

`mix check` behaviour can be customised through `.check.exs`.
For more information see the gh repo:
https://github.com/karolsluszniak/ex_check