fix: update git log command to use pull request SHA
Currently this is not working on Ash Auth and Ash Auth Phoenix because the last commit is actual a synthetic GitHub commit which is the merge of the branch head and main, not the commit we care about checking.

So we need to explicitly specify the SHA representing the head of the branch being merged.

Learn more about how this works:

https://www.kenmuse.com/blog/the-many-shas-of-a-github-pull-request/

https://www.baeldung.com/ops/github-actions-commit-sha